added REST api

+++ src/main/java/com/example/solidbankingapp/service/impl/AccountDepositServiceImpl.java
@@ -6,18 +6,28 @@ import com.example.solidbankingapp.service.AccountDepositService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import java.util.Objects;
+
 @Component
 public class AccountDepositServiceImpl implements AccountDepositService {
-    @Autowired
     private AccountDAO accountDAO;
 
+    @Autowired
+    public AccountDepositServiceImpl(AccountDAO accountDAO) {
+        this.accountDAO = accountDAO;
+    }
+
     @Override
-    public void deposit(double amount, Account account) {
+    public String deposit(double amount, Account account) throws NullPointerException {
         double oldBalance = account.getBalance();
         double newBalance = amount + oldBalance;
+        System.out.println(account);
+        if (!accountDAO.existsById(account.getId())){
+            return "Account does not exist";
+        }
         accountDAO.deleteById(account.getId());
         account.setBalance(newBalance);
         accountDAO.save(account);
-        System.out.println(amount + "$ transferred to " + account.getId() + " account");
+        return amount + "$ transferred to " + account.getId() + " account";
     }
 }
